+++ School/OOP_CP_Test/src/Mercedes.java
@@ -8,21 +8,13 @@ public class Mercedes extends Car{
     private String Name;
     private CarColor Color;
 
-    public CarColor getColor() {
-        return Color;
-    }
-    public void setColor(CarColor color) {
-        Color = color;
-    }
+    public CarColor getColor() {return Color;}
+    public void setColor(CarColor color) {Color = color;}
 
-    public String getName() {
-        return Name;
-    }
-    public void setName(String name) {
-        Name = name;
-    }
+    public String getName() {return Name;}
+    public void setName(String name) {Name = name;}
 
-    public Mercedes(int speed, int acceleration, int topSpeed) {
+    public Mercedes(float speed, float acceleration, float topSpeed) {
         super(speed, acceleration, topSpeed);
     }
 }
